renamed to RNLocalizeExampleApp from RNLocalizeSampleApp to better align with main repro example app name
