+++ example/windows/RNLocalizeExampleApp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29215.179
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RNLocalizeSampleApp", "RNLocalizeSampleApp\RNLocalizeSampleApp.vcxproj", "{EEC382FF-948D-444C-8318-03380E8A3088}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RNLocalizeExampleApp", "RNLocalizeExampleApp\RNLocalizeExampleApp.vcxproj", "{EEC382FF-948D-444C-8318-03380E8A3088}"
 	ProjectSection(ProjectDependencies) = postProject
 		{F7D32BD0-2749-483E-9A0D-1635EF7E3136} = {F7D32BD0-2749-483E-9A0D-1635EF7E3136}
 	EndProjectSection
