+++ example/windows/RNLocalizeExampleApp/.gitignore
