+++ example/windows/RNLocalizeExampleApp/App.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "App.xaml.g.h"
+
+namespace activation = winrt::Windows::ApplicationModel::Activation;
+
+namespace winrt::RNLocalizeExampleApp::implementation
+{
+struct App : AppT<App>
+{
+  App() noexcept;
+  void OnLaunched(activation::LaunchActivatedEventArgs const &);
+  void OnSuspending(IInspectable const &, Windows::ApplicationModel::SuspendingEventArgs const &);
+  void OnNavigationFailed(IInspectable const &, Windows::UI::Xaml::Navigation::NavigationFailedEventArgs const &);
+
+private:
+  using super = AppT<App>;
+};
+} // namespace winrt::RNLocalizeExampleApp::implementation
