+++ example/windows/RNLocalizeExampleApp/App.idl
@@ -1,0 +1,3 @@
+namespace RNLocalizeExampleApp
+{
+}
