+++ example/windows/RNLocalizeExampleApp/App.xaml
@@ -1,8 +1,8 @@
 ﻿<react:ReactApplication
-    x:Class="RNLocalizeSampleApp.App"
+    x:Class="RNLocalizeExampleApp.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:RNLocalizeSampleApp"
+    xmlns:local="using:RNLocalizeExampleApp"
     xmlns:react="using:Microsoft.ReactNative">
     <Application.Resources>
         <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
