+++ example/windows/RNLocalizeExampleApp/Assets/LockScreenLogo.scale-200.png
