+++ example/windows/RNLocalizeExampleApp/Assets/SplashScreen.scale-200.png
