+++ example/windows/RNLocalizeExampleApp/Assets/Square44x44Logo.scale-200.png
