+++ example/windows/RNLocalizeExampleApp/Assets/StoreLogo.png
