+++ example/windows/RNLocalizeExampleApp/AutolinkedNativeModules.g.cpp
