+++ example/windows/RNLocalizeExampleApp/AutolinkedNativeModules.g.h
