+++ example/windows/RNLocalizeExampleApp/AutolinkedNativeModules.g.props
