+++ example/windows/RNLocalizeExampleApp/MainPage.cpp
@@ -1,0 +1,20 @@
+﻿#include "pch.h"
+#include "MainPage.h"
+#if __has_include("MainPage.g.cpp")
+#include "MainPage.g.cpp"
+#endif
+
+#include "App.h"
+
+using namespace winrt;
+using namespace Windows::UI::Xaml;
+
+namespace winrt::RNLocalizeExampleApp::implementation
+{
+MainPage::MainPage()
+{
+    InitializeComponent();
+    auto app = Application::Current().as<App>();
+    ReactRootView().ReactNativeHost(app->Host());
+}
+} // namespace winrt::RNLocalizeExampleApp::implementation
