+++ example/windows/RNLocalizeExampleApp/MainPage.idl
@@ -1,4 +1,4 @@
-namespace RNLocalizeSampleApp
+namespace RNLocalizeExampleApp
 {
     [default_interface]
     runtimeclass MainPage : Windows.UI.Xaml.Controls.Page
