+++ example/windows/RNLocalizeExampleApp/MainPage.xaml
@@ -1,8 +1,8 @@
 <Page
-    x:Class="RNLocalizeSampleApp.MainPage"
+    x:Class="RNLocalizeExampleApp.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:RNLocalizeSampleApp"
+    xmlns:local="using:RNLocalizeExampleApp"
     xmlns:react="using:Microsoft.ReactNative"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
