+++ example/windows/RNLocalizeExampleApp/Package.appxmanifest
@@ -14,7 +14,7 @@
   <mp:PhoneIdentity PhoneProductId="05d129f1-90c2-4faa-a2c5-b313b15aa271" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
 
   <Properties>
-    <DisplayName>RNLocalizeSampleApp</DisplayName>
+    <DisplayName>RNLocalizeExampleApp</DisplayName>
     <PublisherDisplayName>namro</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -31,12 +31,12 @@
     <Application
       Id="App"
       Executable="$targetnametoken$.exe"
-      EntryPoint="RNLocalizeSampleApp.App">
+      EntryPoint="RNLocalizeExampleApp.App">
       <uap:VisualElements
-        DisplayName="RNLocalizeSampleApp"
+        DisplayName="RNLocalizeExampleApp"
         Square150x150Logo="Assets\Square150x150Logo.png"
         Square44x44Logo="Assets\Square44x44Logo.png"
-        Description="RNLocalizeSampleApp"
+        Description="RNLocalizeExampleApp"
         BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png"/>
         <uap:SplashScreen Image="Assets\SplashScreen.png" />
