+++ example/windows/RNLocalizeExampleApp/PropertySheet.props
