+++ example/windows/RNLocalizeExampleApp/RNLocalizeExampleApp_TemporaryKey.pfx
