+++ example/windows/RNLocalizeExampleApp/ReactPackageProvider.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "winrt/Microsoft.ReactNative.h"
+
+namespace winrt::RNLocalizeExampleApp::implementation
+{
+struct ReactPackageProvider : winrt::implements<ReactPackageProvider, winrt::Microsoft::ReactNative::IReactPackageProvider>
+{
+public: // IReactPackageProvider
+    void CreatePackage(winrt::Microsoft::ReactNative::IReactPackageBuilder const &packageBuilder) noexcept;
+};
+} // namespace winrt::RNLocalizeExampleApp::implementation
