+++ example/windows/RNLocalizeExampleApp/packages.config
