+++ example/windows/RNLocalizeExampleApp/pch.cpp
