+++ example/windows/RNLocalizeExampleApp/pch.h
