--- example/windows/RNLocalizeSampleApp/App.idl
@@ -1,3 +1,0 @@
-namespace RNLocalizeSampleApp
-{
-}
