--- example/windows/RNLocalizeSampleApp/MainPage.cpp
@@ -1,20 +1,0 @@
-﻿#include "pch.h"
-#include "MainPage.h"
-#if __has_include("MainPage.g.cpp")
-#include "MainPage.g.cpp"
-#endif
-
-#include "App.h"
-
-using namespace winrt;
-using namespace Windows::UI::Xaml;
-
-namespace winrt::RNLocalizeSampleApp::implementation
-{
-    MainPage::MainPage()
-    {
-        InitializeComponent();
-        auto app = Application::Current().as<App>();
-        ReactRootView().ReactNativeHost(app->Host());
-    }
-}
