--- example/windows/RNLocalizeSampleApp/ReactPackageProvider.h
@@ -1,13 +1,0 @@
-#pragma once
-
-#include "winrt/Microsoft.ReactNative.h"
-
-namespace winrt::RNLocalizeSampleApp::implementation
-{
-    struct ReactPackageProvider : winrt::implements<ReactPackageProvider, winrt::Microsoft::ReactNative::IReactPackageProvider>
-    {
-    public: // IReactPackageProvider
-        void CreatePackage(winrt::Microsoft::ReactNative::IReactPackageBuilder const &packageBuilder) noexcept;
-    };
-} // namespace winrt::RNLocalizeSampleApp::implementation
-
